LayoutTests/touchadjustment/rotated-node-expected.txt

Test touch adjustment to a non-rectilinear element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Direct Touches
PASS adjustedNode.id is "rotatedDiv"
PASS adjustedNode.id is "rotatedDiv"
PASS adjustedNode.id is "rotatedDiv"
PASS adjustedNode.id is "rotatedDiv"
PASS adjustedNode.id is "rotatedDiv"
PASS adjustedNode.id is "containerDiv"
PASS adjustedNode.id is "containerDiv"
PASS adjustedNode.id is "containerDiv"

Adjusted Touches
PASS adjustedNode.id is "rotatedDiv"
PASS adjustedNode.id is "rotatedDiv"

Near Misses
PASS adjustedNode.id is "containerDiv"
PASS adjustedNode.id is "containerDiv"
PASS successfullyParsed is true

TEST COMPLETE
